Implement Roles / First steps of UI

+++ app/Http/Controllers/UserController.php
@@ -6,23 +6,34 @@ use Illuminate\Http\Request;
 use App\User;
 use App\Http\Resources\User as UserResource;
 use Illuminate\Support\Facades\DB;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
     public function __construct() {
-        $this->middleware("auth:api", ["except" => ["index"]]);
+        $this->middleware("auth:api");
     }
 
     public function index($id, Request $request) {
+        if(!$request->user()->can("view users")) {
+            return reponse(null, 403);
+        }
+
         $user = User::findOrFail($id);
         $user->recommendations->makeHidden(["content"]);
         $user->followersCount = $user->followers()->count();
+        $user->role = $user->roles()->first()->name;
 
         return new UserResource($user);
     }
 
     public function follow(Request $request) {
         $user = $request->user();
+
+        if(!$user->can("follow users")) {
+            return response(null, 403);
+        }
+
         $followUser = User::findOrFail($request->id);
 
         if($user->id != $followUser->id) {
@@ -35,6 +46,11 @@ class UserController extends Controller
 
     public function unfollow(Request $request) {
         $user = $request->user();
+
+        if(!$user->can("follow users")) {
+            return response(null, 403);
+        }
+
         $unfollowUser = User::findOrFail($request->id);
 
         if($user->follows()->detach($unfollowUser)) {
@@ -46,6 +62,11 @@ class UserController extends Controller
 
     public function follows($id, Request $request) {
         $user = $request->user();
+
+        if(!$user->can("view users")) {
+            return response(null, 403);
+        }
+
         $followsUser = User::findOrFail($id);
         $isFollowing = !!DB::table("followers")
             ->select("user_from_id", "user_to_id")
@@ -56,5 +77,50 @@ class UserController extends Controller
             ->count();
 
         return response()->json($isFollowing);
+    }
+
+    public function update(Request $request) {
+        $user_request = $request->user();
+        $user = User::findOrFail($request->id);
+
+        if(!$user_request->can("update any user")) {
+            return response(null, 403);
+        }
+
+        $user->username = $request->username;
+        $user->email = $request->email;
+        $user->links = $request->links;
+        $user->biography = $request->biography;
+
+        return $user->save();
+    }
+
+    public function destroy($id, Request $request) {        
+        if(!$request->user()->can("delete any user")) {
+            return response(null, 403);
+        }
+
+        $user = User::findOrFail($id);
+        if($user->delete()) {
+            return new UserResource($user);
+        }
+    }
+
+    public function role(Request $request) {
+        if(!$request->user()->can("update any user")) {
+            return response(null, 403);
+        }
+
+        $user = User::findOrFail($request->user_id);
+        $role = Role::findOrFail($request->role_id);
+        $user->syncRoles($role);
+    }
+
+    public function roles(Request $request) {
+        if($request->user()->can("view roles")) {
+            return Role::all();
+        } else {
+            return response(null, 403);
+        }
     }
 }
